requirements.txt
tensorflow-macos>=2.12.0
tensorflow-metal>=1.0.0
torch>=1.12.0
torchvision>=0.13.0
pandas
matplotlib
h5py
numpy
scikit-learn
